requirements.txt
cffi==1.17.1
colorama==0.4.6
cryptography==43.0.1
iniconfig==2.0.0
numpy==2.1.1
oracledb==2.4.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1